ClipboardCopyLink コンポーネントを削除し、RoomInfo コンポーネントにコピー機能を統合
